requirements.txt
ansible
bcrypt
celery[redis]
eventlet
flask-collect
flask-cors
flask-jwt-extended
flask-mongoengine
flask-rest-api
flask-security
flask-socketio==3.0.2
gunicorn
marshmallow
pynacl<1.3
